corrected base_controller namespace
